wip: panel-header: More merge conflicts during cherry pick
