+++ public/app/plugins/panel/graph2/module.tsx
@@ -75,5 +75,4 @@ export class GraphOptions extends PureComponent<PanelOptionsProps<Options>> {
 }
 
 export { Graph2 as PanelComponent, GraphOptions as PanelOptionsComponent };
-export { withMenuOptions } from './withMenuOptions';
 export { moduleMenu } from './moduleMenu';
