+++ public/app/types/panel.ts
@@ -12,3 +12,26 @@ export interface PanelOptionsProps<T = any> {
   options: T;
   onChange: (options: T) => void;
 }
+
+export enum PanelHeaderMenuItemTypes { // TODO: Evaluate. Remove?
+  Button = 'Button', // ?
+  Divider = 'Divider',
+  Link = 'Link',
+  SubMenu = 'SubMenu',
+}
+
+export interface PanelHeaderMenuItemProps {
+  type: PanelHeaderMenuItemTypes;
+  text?: string;
+  iconClassName?: string;
+  handleClick?: () => void;
+  shortcut?: string;
+  children?: any;
+  subMenu?: PanelHeaderMenuItemProps[];
+  role?: string;
+}
+
+export interface PanelMenuExtras {
+  additionalMenuItems: PanelHeaderMenuItemProps[];
+  additionalSubMenuItems: PanelHeaderMenuItemProps[];
+}
