+++ public/app/types/plugins.ts
@@ -13,7 +13,6 @@ export interface PluginExports {
   PanelCtrl?;
   PanelComponent?: ComponentClass<PanelProps>;
   PanelOptionsComponent: ComponentClass<PanelOptionsProps>;
-  withMenuOptions?: any;
   moduleMenu?: any;
 }
